Moved the calendar subcommand in the command view
